refactor(multiverse): Move the room view into a separate widget
