+++ labs/multiverse/src/widgets/mod.rs
@@ -4,4 +4,5 @@ pub mod linked_chunk;
 pub mod read_receipts;
 pub mod recovery;
 pub mod room_list;
+pub mod room_view;
 pub mod status;
